型変更とポート番号変更、フロントの残り部分の追加
fix:Long型の変数をInteger型に変更
change:HostURLをSpringBootでフォルトポートに変更
add:客側画面の追加

+++ demo/src/main/java/com/example/demo/dto/ItemRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 @Data
 public class ItemRequest {
 	/** 商品番号 */
-	private Long itemId; //主キー
+	private Integer itemId; //主キー
 	/** 商品名 */
 	private String itemName;
 	/** 単価 */
