+++ demo/src/main/java/com/example/demo/entity/OrderTable.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class OrderTable { // <-DB用
 	/** 注文番号 */
-	private Long orderId;
+	private Integer orderId;
 	/** 注文日時 */
 	private LocalDateTime orderDate;
 	/** 予約時間 */
@@ -14,5 +14,5 @@ public class OrderTable { // <-DB用
 	/** 受け渡しの状態(0,1,2) */
 	private Integer servingStatus;
 	/** ユーザID */
-	private Long userId;
+	private Integer userId;
 }
