+++ demo/src/main/java/com/example/demo/model/Item.java
@@ -4,7 +4,7 @@ import lombok.Data;
 @Data
 public class Item { //事前登録型　登録の詳細はitemMemo参照
 	/** 商品番号 */
-	private Long itemId; //主キー
+	private Integer itemId; //主キー
 	/** 商品名 */
 	private String itemName;
 	/** 単価 */
